requirements.txt
Flask~=1.1.1
Werkzeug~=1.0.0
WTForms~=2.2.1

python-dotenv~=0.17.0